Fix race condition, dead code, and correctness issues from code review
Critical fix: Remove global activeNamedGraphs variable. Thread
namedGraphs as parameter through evalPattern, evalPatternWithBindings,
evalGraphPattern, and evalExprWithGraph. This makes concurrent query
evaluation safe.

Dead code: Remove unused bindingKeyNorm, bindingKey functions.
Replace manual insertion sort (sortStrings) with slices.Sort.

Correctness: Use int64 instead of int for integer arithmetic results
to prevent overflow. Remove duplicate base URI setup in EvalQuery.
Fix redundant prefix validation condition. Run gofmt on all files.

+++ sparql/functions.go
@@ -750,7 +750,7 @@ func castXSD(name string, val rdflibgo.Term) rdflibgo.Term {
 		if isNumericDatatype(dt) {
 			// From numeric: truncate to integer
 			if f, err := strconv.ParseFloat(s, 64); err == nil {
-				return rdflibgo.NewLiteral(int(f), rdflibgo.WithDatatype(rdflibgo.XSDInteger))
+				return rdflibgo.NewLiteral(int64(f), rdflibgo.WithDatatype(rdflibgo.XSDInteger))
 			}
 		}
 		// From string/plain: must be a valid integer lexical form
